Data from a Simulated Motorcycle Accident

A data frame giving a series of measurements of head acceleration in a simulated motorcycle accident, used to test crash helmets.

Usage

mcycle_train

mcycle_test

Format

An object of class data.frame with 89 rows and 2 columns.

An object of class data.frame with 44 rows and 2 columns.

Source

Silverman, B. W. (1985) Some aspects of the spline smoothing approach to non-parametric curve fitting. Journal of the Royal Statistical Society series B 47, 1--52.

Details

This data is originally available as MASS::mcycle.

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S-PLUS. Fourth Edition. Springer.

Examples

mcycle_train
#>     times  accel
#> 1     2.4    0.0
#> 2     2.6   -1.3
#> 4     3.6    0.0
#> 5     4.0   -2.7
#> 7     6.6   -2.7
#> 8     6.8   -1.3
#> 10    8.2   -2.7
#> 11    8.8   -1.3
#> 13    9.6   -2.7
#> 14   10.0   -2.7
#> 16   10.6   -2.7
#> 17   11.0   -5.4
#> 19   13.2   -2.7
#> 20   13.6   -2.7
#> 22   14.6  -13.3
#> 23   14.6   -5.4
#> 25   14.6   -9.3
#> 26   14.6  -16.0
#> 28   14.8   -2.7
#> 29   15.4  -22.8
#> 31   15.4  -53.5
#> 32   15.4  -54.9
#> 34   15.6  -21.5
#> 35   15.8  -21.5
#> 37   16.0  -42.9
#> 38   16.0  -26.8
#> 40   16.2  -50.8
#> 41   16.2  -61.7
#> 43   16.4  -80.4
#> 44   16.6  -59.0
#> 46   16.8  -91.1
#> 47   16.8  -77.7
#> 49   17.6  -85.6
#> 50   17.6 -123.1
#> 52   17.8  -99.1
#> 53   17.8 -104.4
#> 55   18.6  -50.8
#> 56   19.2 -123.1
#> 58   19.4  -72.3
#> 59   19.6 -127.2
#> 61   20.4 -117.9
#> 62   21.2 -134.0
#> 64   21.8 -108.4
#> 65   22.0 -123.1
#> 67   23.4 -128.5
#> 68   24.0 -112.5
#> 70   24.2  -81.8
#> 71   24.6  -53.5
#> 73   25.0  -57.6
#> 74   25.4  -72.3
#> 76   25.6  -26.8
#> 77   26.0   -5.4
#> 79   26.2  -21.5
#> 80   26.4  -65.6
#> 82   27.2  -45.6
#> 83   27.2  -24.2
#> 85   27.6    4.0
#> 86   28.2   12.0
#> 88   28.4   37.5
#> 89   28.6   46.9
#> 91   30.2   36.2
#> 92   31.0   75.0
#> 94   32.0   54.9
#> 95   32.0   48.2
#> 97   33.4   16.0
#> 98   33.8   45.6
#> 100  34.8   75.0
#> 101  35.2  -16.0
#> 103  35.4   69.6
#> 104  35.6   34.8
#> 106  36.2  -37.5
#> 107  36.2   22.8
#> 109  38.0   10.7
#> 110  39.2    5.4
#> 112  40.0  -21.5
#> 113  40.4  -13.3
#> 115  41.6  -10.7
#> 116  42.4   29.4
#> 118  42.8  -10.7
#> 119  43.0   14.7
#> 121  44.4    0.0
#> 122  45.0   10.7
#> 124  47.8  -26.8
#> 125  47.8  -14.7
#> 127  50.6    0.0
#> 128  52.0   10.7
#> 130  55.0   -2.7
#> 131  55.0   10.7
#> 133  57.6   10.7

mcycle_test
#>     times  accel
#> 3     3.2   -2.7
#> 6     6.2   -2.7
#> 9     7.8   -2.7
#> 12    8.8   -2.7
#> 15   10.2   -5.4
#> 18   11.4    0.0
#> 21   13.8    0.0
#> 24   14.6   -5.4
#> 27   14.6  -22.8
#> 30   15.4  -32.1
#> 33   15.6  -40.2
#> 36   15.8  -50.8
#> 39   16.2  -21.5
#> 42   16.4   -5.4
#> 45   16.8  -71.0
#> 48   17.6  -37.5
#> 51   17.6 -101.9
#> 54   18.6 -112.5
#> 57   19.4  -85.6
#> 60   20.2 -123.1
#> 63   21.4 -101.9
#> 66   23.2 -123.1
#> 69   24.2  -95.1
#> 72   25.0  -64.4
#> 75   25.4  -44.3
#> 78   26.2 -107.1
#> 81   27.0  -16.0
#> 84   27.2    9.5
#> 87   28.4  -21.5
#> 90   29.4  -17.4
#> 93   31.2    8.1
#> 96   32.8   46.9
#> 99   34.4    1.3
#> 102  35.2  -54.9
#> 105  35.6   32.1
#> 108  38.0   46.9
#> 111  39.4   -1.3
#> 114  41.6   30.8
#> 117  42.8    0.0
#> 120  44.0   -1.3
#> 123  46.6   10.7
#> 126  48.8  -13.3
#> 129  53.2  -14.7
#> 132  55.4   -2.7