Economic mobility dataset

Data for discussing the probability of improving one's economic situation across 741 Metropolitan Statistical Areas in the US.

Usage

mobility

Format

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 741 rows and 43 columns.

Source

This data set was based on the paper "Where is the Land of Opportunity? The Geography of Intergenerational Mobility in the United States" (Raj Chetty, Nathaniel Hendren, Patrick Kline, Emmanuel Saez, QJE 2014), and its replication files. Both the paper and the data sets are available from https://opportunityinsights.org.

Details

The variables are:

Variable	Description

Mobility	The probability that a child born in 1980–1982 into the lowest quintile (20%) of household income will be in the top quintile at age 30. Individuals are assigned to the community they grew up in, not the one they were in as adults.

Population	in 2000.

Urban	Is the community primarily urban or rural?

Black	percentage of individuals who marked black (and nothing else) on census forms.

Racial Segregation	a measure of residential segregation by race.

Income Segregation	Similarly but for income.

Segregation of poverty	Specifically a measure of residential segregation for those in the bottom quarter of the national income distribution.

Segregation of affluence	Residential segregation for those in the top quarter.

Commute	Fraction of workers with a commute of less than 15 minutes.

Mean income	Average income per capita in 2000.

Gini	A measure of income inequality, which would be 0 if all incomes were perfectly equal, and tends towards 100 as all the income is concentrated among the richest individuals (see Wikipedia, s.v. "Gini coefficient").

Share 1%	Share of the total income of a community going to its richest 1%.

Gini bottom 99%	Gini coefficient among the lower 99% of that community.

Fraction middle class	Fraction of parents whose income is between the national 25th and 75th percentiles.

Local tax rate	Fraction of all income going to local taxes.

Local government spending	per capita.

Progressivity	Measure of how much state income tax rates increase with income.

EITC	Measure of how much the state contributed to the Earned Income Tax Credit (a sort of negative income tax for very low-paid wage earners).

School expenditures	Average spending per pupil in public schools.

Student/teacher ratio	Number of students in public schools divided by number of teachers.

Test scores	Residuals from a linear regression of mean math and English test scores on household income per capita.

High school dropout rate	Residuals from a linear regression of the dropout rate on per-capita income.

Colleges per capita

College tuition	in-state, for full-time students

College graduation rate	Again, residuals from a linear regression of the actual graduation rate on household income per capita.

Labor force participation	Fraction of adults in the workforce.

Manufacturing	Fraction of workers in manufacturing.

Chinese imports	Growth rate in imports from China per worker between 1990 and 2000.

Teenage labor	fraction of those age 14–16 who were in the labor force.

Migration in	Migration into the community from elsewhere, as a fraction of 2000 population.

Migration out	Ditto for migration into other communities.

Foreign	fraction of residents born outside the US.

Social capital	Index combining voter turnout, participation in the census, and participation in community organizations.

Religious	Share of the population claiming to belong to an organized religious body.

Violent crime	Arrests per person per year for violent crimes.

Single motherhood	Number of single female households with children divided by the total number of households with children.

Divorced	Fraction of adults who are divorced.

Married	Ditto.

Longitude	Geographic coordinate for the center of the community

Latitude	Ditto

ID	A numerical code, identifying the community.

Name	the name of principal city or town.

State	the state of the principal city or town of the community.

Examples

mobility
#> # A tibble: 741 × 43
#>       ID Name        Mobility State Population Urban Black Seg_racial Seg_income
#>    <dbl> <chr>          <dbl> <chr>      <dbl> <dbl> <dbl>      <dbl>      <dbl>
#>  1   100 Johnson Ci…   0.0622 TN        576081     1 0.021      0.09       0.035
#>  2   200 Morristown    0.0537 TN        227816     1 0.02       0.093      0.026
#>  3   301 Middlesbor…   0.0726 TN         66708     0 0.015      0.064      0.024
#>  4   302 Knoxville     0.0563 TN        727600     1 0.056      0.21       0.092
#>  5   401 Winston-Sa…   0.0448 NC        493180     1 0.174      0.262      0.072
#>  6   402 Martinsvil…   0.0518 VA         92753     0 0.224      0.137      0.024
#>  7   500 Greensboro    0.0474 NC       1055133     1 0.218      0.22       0.068
#>  8   601 North Wilk…   0.0517 NC         90016     0 0.032      0.114      0.012
#>  9   602 Galax         0.0796 VA         64676     0 0.029      0.131      0.005
#> 10   700 Spartanburg   0.0431 SC        354533     1 0.207      0.139      0.045
#> # ℹ 731 more rows
#> # ℹ 34 more variables: Seg_poverty <dbl>, Seg_affluence <dbl>, Commute <dbl>,
#> #   Income <dbl>, Gini <dbl>, Share01 <dbl>, Gini_99 <dbl>, Middle_class <dbl>,
#> #   Local_tax_rate <dbl>, Local_gov_spending <dbl>, Progressivity <dbl>,
#> #   EITC <dbl>, School_spending <dbl>, Student_teacher_ratio <dbl>,
#> #   Test_scores <dbl>, HS_dropout <dbl>, Colleges <dbl>, Tuition <dbl>,
#> #   Graduation <dbl>, Labor_force_participation <dbl>, Manufacturing <dbl>, …