Spotify data for 3 artists

Audio features for all recordings by Taylor Swift, Radiohead, and The National, as available in the Spotify API.

Usage

popmusic_train

popmusic_test

Format

Audio features for one track

artist

Who is performing this music? This data contains tracks by Taylor Swift, Radiohead, and The National.

acousticness

A confidence measure from 0.0 to 1.0 of whether the track is acoustic. 1.0 represents high confidence the track is acoustic.

danceability

Danceability describes how suitable a track is for dancing based on a combination of musical elements including tempo, rhythm stability, beat strength, and overall regularity. A value of 0.0 is least danceable and 1.0 is most danceable.

duration_ms

The duration of the track in milliseconds.

energy

Energy is a measure from 0.0 to 1.0 and represents a perceptual measure of intensity and activity. Typically, energetic tracks feel fast, loud, and noisy. For example, death metal has high energy, while a Bach prelude scores low on the scale. Perceptual features contributing to this attribute include dynamic range, perceived loudness, timbre, onset rate, and general entropy.

instrumentalness

Predicts whether a track contains no vocals. "Ooh" and "aah" sounds are treated as instrumental in this context. Rap or spoken word tracks are clearly "vocal". The closer the instrumentalness value is to 1.0, the greater likelihood the track contains no vocal content. Values above 0.5 are intended to represent instrumental tracks, but confidence is higher as the value approaches 1.0.

key

The key the track is in. Integers map to pitches using standard Pitch Class notation. E.g. 0 = C, 1 = C♯/D♭, 2 = D, and so on. If no key was detected, the value is -1.

liveness

Detects the presence of an audience in the recording. Higher liveness values represent an increased probability that the track was performed live. A value above 0.8 provides strong likelihood that the track is live.

loudness

The overall loudness of a track in decibels (dB). Loudness values are averaged across the entire track and are useful for comparing relative loudness of tracks. Loudness is the quality of a sound that is the primary psychological correlate of physical strength (amplitude). Values typically range between -60 and 0 db.

mode

Mode indicates the modality (major or minor) of a track, the type of scale from which its melodic content is derived. Major is represented by 1 and minor is 0.

speechiness

Speechiness detects the presence of spoken words in a track. The more exclusively speech-like the recording (e.g. talk show, audio book, poetry), the closer to 1.0 the attribute value. Values above 0.66 describe tracks that are probably made entirely of spoken words. Values between 0.33 and 0.66 describe tracks that may contain both music and speech, either in sections or layered, including such cases as rap music. Values below 0.33 most likely represent music and other non-speech-like tracks.

tempo

The overall estimated tempo of a track in beats per minute (BPM). In musical terminology, tempo is the speed or pace of a given piece and derives directly from the average beat duration.

time_signature

An estimated time signature. The time signature (meter) is a notational convention to specify how many beats are in each bar (or measure). The time signature ranges from 3 to 7 indicating time signatures of "3/4", to "7/4".

valence

A measure from 0.0 to 1.0 describing the musical positiveness conveyed by a track. Tracks with high valence sound more positive (e.g. happy, cheerful, euphoric), while tracks with low valence sound more negative (e.g. sad, depressed, angry).

explicit

Are there bad words that prevent typical play on the radio (not safe for work).

An object of class tbl_df (inherits from tbl, data.frame) with 1269 rows and 15 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 425 rows and 15 columns.

Source

Rohan Alexander (2023). Telling Stories with Data: With Applications in R. CRC Press, Toronto. https://tellingstorieswithdata.com/.

The example data are available with the textbook, originally from the Spotify API and described in Chapter 7.

Documentation is from Spotify API.

Details

popmusic_train is the training set.

popmusic_test is the test set. It contains a held out set of 425 songs.

Examples

popmusic_train
#> # A tibble: 1,269 × 15
#>    artist      danceability energy   key loudness  mode speechiness acousticness
#>    <fct>              <dbl>  <dbl> <int>    <dbl> <int>       <dbl>        <dbl>
#>  1 Taylor Swi…        0.781  0.357     0   -16.4      1      0.912       0.717  
#>  2 Taylor Swi…        0.627  0.266     9   -15.4      1      0.929       0.796  
#>  3 Taylor Swi…        0.516  0.917    11    -3.19     0      0.0827      0.0139 
#>  4 Taylor Swi…        0.629  0.757     1    -8.37     0      0.0512      0.00384
#>  5 Taylor Swi…        0.686  0.705     9   -10.8      1      0.249       0.832  
#>  6 Taylor Swi…        0.522  0.691     2    -4.82     1      0.0307      0.00609
#>  7 Taylor Swi…        0.31   0.374     6    -8.46     1      0.0275      0.761  
#>  8 Taylor Swi…        0.705  0.621     2    -8.09     1      0.0334      0.101  
#>  9 Taylor Swi…        0.553  0.604     1    -5.30     0      0.0258      0.202  
#> 10 Taylor Swi…        0.419  0.908     9    -5.16     1      0.0651      0.00048
#> # ℹ 1,259 more rows
#> # ℹ 7 more variables: instrumentalness <dbl>, liveness <dbl>, valence <dbl>,
#> #   tempo <dbl>, time_signature <int>, duration_ms <int>, explicit <lgl>

popmusic_test
#> # A tibble: 425 × 15
#>    artist      danceability energy   key loudness  mode speechiness acousticness
#>    <fct>              <dbl>  <dbl> <int>    <dbl> <int>       <dbl>        <dbl>
#>  1 Taylor Swi…        0.694  0.38      2   -10.3      1      0.0614       0.416 
#>  2 Taylor Swi…        0.7    0.55      7    -9.13     1      0.0653       0.0661
#>  3 Taylor Swi…        0.334  0.161     0   -14.9      1      0.0506       0.967 
#>  4 Taylor Swi…        0.483  0.84      7    -6.51     1      0.119        0.43  
#>  5 Taylor Swi…        0.65   0.404     7    -8.40     1      0.0356       0.0616
#>  6 Taylor Swi…        0.722  0.494     7    -8.69     1      0.204        0.216 
#>  7 Taylor Swi…        0.654  0.638     8    -7.08     1      0.075        0.0727
#>  8 Taylor Swi…        0.696  0.485     7   -10.6      1      0.123        0.103 
#>  9 Taylor Swi…        0.571  0.477    11    -9.74     0      0.246        0.313 
#> 10 Taylor Swi…        0.681  0.396     0   -12.1      1      0.0487       0.487 
#> # ℹ 415 more rows
#> # ℹ 7 more variables: instrumentalness <dbl>, liveness <dbl>, valence <dbl>,
#> #   tempo <dbl>, time_signature <int>, duration_ms <int>, explicit <lgl>